--- ino/MySat01/ina.h
@@ -1,21 +1,0 @@
-void printValuesINA219()
-{
-  shuntvoltage = ina219.getShuntVoltage_mV();
-  busvoltage = ina219.getBusVoltage_V();
-  current = ina219.getCurrent_mA();
-  power = ina219.getPower_mW();
-  loadvoltage = busvoltage + (shuntvoltage / 1000);
-
-  Serial.print("BV"); Serial.print("\t"); // Bus Voltage
-  Serial.print("SV"); Serial.print("\t"); // Shunt Voltage
-  Serial.print("LV"); Serial.print("\t"); // Load Voltage
-  Serial.print("C"); Serial.print("\t");  // Current
-  Serial.println("P");  // Power
-
-  Serial.print(busvoltage); Serial.print("\t"); 
-  Serial.print(shuntvoltage); Serial.print("\t");
-  Serial.print(loadvoltage); Serial.print("\t");
-  Serial.print(current); Serial.print("\t");
-  Serial.println(power);
-
-}
