--- ino/MySat01/mcu.h
@@ -1,46 +1,0 @@
-void printValues9250() {
-    uint8_t sensorId = 0x69;
-  if (mySensor.readId(&sensorId) == 0) {
-    Serial.println("sensorId: " + String(sensorId));
-  } else {
-    Serial.println("Cannot read sensorId");
-  }
- 
-  if (mySensor.accelUpdate() == 0) {
-    aX = mySensor.accelX();
-    aY = mySensor.accelY();
-    aZ = mySensor.accelZ();
-    aSqrt = mySensor.accelSqrt();
-    Serial.println("accelX: " + String(aX));
-    Serial.println("accelY: " + String(aY));
-    Serial.println("accelZ: " + String(aZ));
-    Serial.println("accelSqrt: " + String(aSqrt));
-  } else {
-    Serial.println("Cannod read accel values");
-  }
- 
-  if (mySensor.gyroUpdate() == 0) {
-    gX = mySensor.gyroX();
-    gY = mySensor.gyroY();
-    gZ = mySensor.gyroZ();
-    Serial.println("gyroX: " + String(gX));
-    Serial.println("gyroY: " + String(gY));
-    Serial.println("gyroZ: " + String(gZ));
-  } else {
-    Serial.println("Cannot read gyro values");
-  }
- 
-  if (mySensor.magUpdate() == 0) {
-    mX = mySensor.magX();
-    mY = mySensor.magY();
-    mZ = mySensor.magZ();
-    mDirection = mySensor.magHorizDirection();
-    Serial.println("magX: " + String(mX));
-    Serial.println("maxY: " + String(mY));
-    Serial.println("magZ: " + String(mZ));
-    Serial.println("horizontal direction: " + String(mDirection));
-  } else {
-    Serial.println("Cannot read mag values");
-  }
-
-}
