--- ino/MySat01/sun.h
@@ -1,22 +1,0 @@
-
-void printValuesPhotoRes(){
-
-Serial.println("Phores"); 
-
- Serial.print("Photo 1: ");
-// ph1 = analogRead(12);
- Serial.println(ph1);
-
- Serial.print("Photo 2: ");
-// ph2 = analogRead(13);
- Serial.println(ph2);
-
- Serial.print("Photo 3: ");
-// ph3 = analogRead(14);
- Serial.println(ph3);
-
- Serial.print("Photo 4: ");
-// ph4 = analogRead(15);
- Serial.println(ph4);
-
-}
